+++ tidy3d/components/types.py
@@ -200,6 +200,7 @@ PlanePosition = Literal["bottom", "middle", "top"]
 ClipOperationType = Literal["union", "intersection", "difference", "symmetric_difference"]
 BoxSurface = Literal["x-", "x+", "y-", "y+", "z-", "z+"]
 LengthUnit = Literal["nm", "μm", "um", "mm", "cm", "m"]
+PriorityMode = Literal["equal", "conductor"]
 
 """ medium """